requirements.txt
mcrcon==0.7.0
py-cord==2.6.0
python-dotenv==1.0.1
tortoise-orm==0.21.6